third_party/WebKit/LayoutTests/media/track/track-webvtt-tc027-empty-cue.html

<!DOCTYPE html>
<title>Tests that unsupported markup is properly ignored.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// TODO(srirama.m): Correct the title and vtt file description.
check_cues_from_track("captions-webvtt/tc027-empty-cue.vtt", function(track) {
    assert_equals(track.cues.length, 3);
});
</script>
